refactor/c: split C lang module into submodules
